LayoutTests/fast/dom/isEqualNode-after-removeAttribute.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>

description("Tests that Node.isEqualNode behaves correctly after an element that originally had attributes has had them removed.");

e1 = document.createElement('span');
e2 = document.createElement('span');

shouldBe("e1.isEqualNode(e2)", "true");
e2.setAttribute("foo", "bar");
shouldBe("e1.isEqualNode(e2)", "false");
e2.removeAttribute("foo");
shouldBe("e1.isEqualNode(e2)", "true");

e1 = document.createElement('span');
e2 = document.createElement('span');

shouldBe("e1.isEqualNode(e2)", "true");
e1.setAttribute("foo", "bar");
shouldBe("e1.isEqualNode(e2)", "false");
e1.removeAttribute("foo");
shouldBe("e1.isEqualNode(e2)", "true");

</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
