third_party/WebKit/LayoutTests/fast/dom/isEqualNode-after-removeAttribute-expected.txt

Tests that Node.isEqualNode behaves correctly after an element that originally had attributes has had them removed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS e1.isEqualNode(e2) is true
PASS e1.isEqualNode(e2) is false
PASS e1.isEqualNode(e2) is true
PASS e1.isEqualNode(e2) is true
PASS e1.isEqualNode(e2) is false
PASS e1.isEqualNode(e2) is true
PASS successfullyParsed is true

TEST COMPLETE
